HP bar 추가

+++ DuckyDoggyClient/Mesh.cpp
@@ -841,7 +841,7 @@ CWaterMesh::~CWaterMesh()
 }
 
 
-CUIMesh::CUIMesh(ID3D12Device * pd3dDevice, ID3D12GraphicsCommandList * pd3dCommandList, int nWidth, int nLength)
+CUIMesh::CUIMesh(ID3D12Device * pd3dDevice, ID3D12GraphicsCommandList * pd3dCommandList, float nWidth, float nLength)
 {
 	m_nVertices = 4;
 	m_d3dPrimitiveTopology = D3D_PRIMITIVE_TOPOLOGY_TRIANGLESTRIP;
